refactor(check): T7a split check_completeness into per-rule family modules (F30)
